Module description may be incorrect
I've just started playing with Flow and reading the manual so forgive me if this is incorrect. It seems to me (and Flow's Modules menu seems to back me up) that modules with light blue title bars are for modifying partials, not modulation sources.

Very strangely, when I first wrote this pull request it was at the same time you were updating the manual to v9. When I created the change, it suddenly showed a whole lot of changes that I didn't make. 😯 After I figured out what was going on I deleted my fork and started over.